updated docker name and included to-do in scripts
