+++ docker-compose.yml
@@ -11,7 +11,7 @@ services:
       - app-tier
     ports:
       - 7000:7000
-      - 9042:9042
+      - 9042:9042 #db port
     volumes:
       - .:/bitnami
     environment:
@@ -26,7 +26,7 @@ services:
       - app-tier
     ports:
       - 4040:4040
-      - 8888:8888
+      - 8888:8888 #Jupyter lab port
     volumes:
       - ./assignment:/home/jovyan/work
 
@@ -36,7 +36,7 @@ services:
     networks:
       - app-tier
     ports:
-      - 8000:8000
+      - 8000:8000 #default swagger ui api port
       - 8080:8080
     volumes:
       - ./fastapi-app:/app
